Add files via upload
This is env.yml file which is used here to store environment variables or configuration settings in a YAML format. 
The host for this is localhost:3306 and base_url is "http//localhost:8080/"
I have used MySQL database namely production and to connect with the database there is a specific credentials user is "root" and password used is,"prashantpatel".